# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Intuition Core Query API
  version: 0.1.0
  description: |
    REST API for reading and writing the local Intuition knowledge graph.

    Authentication is controlled by `API_AUTH`:
    - `public-read` (default): read endpoints are open; write endpoints require a writable bearer API key.
    - `gated`: all endpoints require a valid bearer API key except `/health` and `/api/schema`.
    - `open`: no endpoint requires a key; writes are unattributed.
servers:
  - url: http://localhost:3000
    description: Local Docker Compose stack
security: []
...
paths:
...
  /api/schema:
    get:
      tags: [Schema]
      summary: Get machine-readable KG table metadata
      description: Returns cached metadata for base tables in the `kg` schema only.
      responses:
        "200":
          description: KG schema metadata.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/KgSchemaMetadata"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/atoms:
    get:
      tags: [Atoms]
      summary: List public atoms
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: classification_type
          in: query
          schema:
            type: string
          description: Filter by atom classification type.
        - name: q
          in: query
          schema:
            type: string
          description: Substring search over atom search text.
      responses:
        "200":
          description: Public atoms ordered newest first.
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AtomListItem"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "401":
          $ref: "#/components/responses/ApiKeyRequired"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Atoms]
      summary: Create an atom
      description: Creates or finds the deterministic atom for a string, URL, or JSON input.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [input]
              properties:
                input:
                  type: string
                  minLength: 1
                  maxLength: 64000
            examples:
              githubRepo:
                value:
                  input: https://github.com/oven-sh/bun
      responses:
        "200":
          description: Atom already existed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateAtomResponse"
        "201":
          description: Atom created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateAtomResponse"
              examples:
                created:
                  value:
                    data:
                      id: "0x951d18ba45fdf5ef5c3304ea3ae46a45e00b68464e2b1561cce4b48ffa1c09d3"
                      created: true
                      createdBy: "0x1111111111111111111111111111111111111111"
        "400":
          description: Invalid JSON or invalid input.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/ApiKeyRequired"
        "403":
          $ref: "#/components/responses/ApiKeyNotWritable"
        "413":
          description: Input is too large.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/atoms/{id}:
    get:
      tags: [Atoms]
      summary: Get one public atom
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/TermId"
      responses:
        "200":
          description: Atom row.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Atom"
        "401":
          $ref: "#/components/responses/ApiKeyRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/atoms/{id}/triples:
    get:
      tags: [Atoms, Triples]
      summary: List public triples touching an atom
      description: Matches the atom id in subject, predicate, or object position.
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/TermId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/ExpandTerms"
      responses:
        "200":
          description: Triples touching the atom.
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Triple"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "401":
          $ref: "#/components/responses/ApiKeyRequired"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/atoms/{id}/artifacts:
    get:
      tags: [Atoms]
      summary: List enrichment artifacts attached to an atom
      description: >-
        One row per provider output (opengraph, github-repo, ...). Newest
        first. 404 when the atom is absent or not public.
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/TermId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Artifacts for the atom.
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Artifact"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "401":
          $ref: "#/components/responses/ApiKeyRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/triples:
    get:
      tags: [Triples]
      summary: List public triples
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/ExpandTerms"
        - name: subject_id
          in: query
          schema:
            type: string
          description: Filter by subject term id.
        - name: predicate_id
          in: query
          schema:
            type: string
          description: Filter by predicate term id.
        - name: object_id
          in: query
          schema:
            type: string
          description: Filter by object term id.
      responses:
        "200":
          description: Public triples ordered newest first.
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Triple"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "401":
          $ref: "#/components/responses/ApiKeyRequired"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Triples]
      summary: Create a triple
      description: Creates or finds the deterministic triple for subject, predicate, and object ids.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subject_id, predicate_id, object_id]
              properties:
                subject_id:
                  type: string
                predicate_id:
                  type: string
                object_id:
                  type: string
            examples:
              claim:
                value:
                  subject_id: "0x951d18ba45fdf5ef5c3304ea3ae46a45e00b68464e2b1561cce4b48ffa1c09d3"
                  predicate_id: "0x0840db4575bf6bdb49b66c21dc40cb4cbb5e1b26bd239d7f56b126c14e452c07"
                  object_id: "0x49ed8bdcd1954e1de3af8bb5fca0f0377ebe598de5dd82ab5b4a45a82d53a53f"
      responses:
        "200":
          description: Triple already existed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTripleResponse"
        "201":
          description: Triple created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTripleResponse"
        "400":
          description: Invalid JSON, invalid input, or invalid term id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/ApiKeyRequired"
        "403":
          $ref: "#/components/responses/ApiKeyNotWritable"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/triples/{id}:
    get:
      tags: [Triples]
      summary: Get one public triple
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/TermId"
        - $ref: "#/components/parameters/ExpandTerms"
      responses:
        "200":
          description: Triple row.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Triple"
        "401":
          $ref: "#/components/responses/ApiKeyRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/events:
    get:
      tags: [Events]
      summary: List graph activity events
      description: >-
        Append-only activity feed (kg.events) — node/triple/predicate/artifact
        creations, onchain and off.
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: entity_kind
          in: query
          schema:
            type: string
            enum: [node, triple, predicate, artifact]
          description: Filter by entity kind.
        - name: event_type
          in: query
          schema:
            type: string
          description: Filter by event type.
        - name: entity_id
          in: query
          schema:
            type: string
          description: Filter by entity id.
      responses:
        "200":
          description: Events, newest first.
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Event"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "401":
          $ref: "#/components/responses/ApiKeyRequired"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/predicates:
    get:
      tags: [Predicates]
      summary: List predicate registry
      security:
        - bearerAuth: []
        - {}
      responses:
        "200":
          description: Seeded predicate registry ordered by slug.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Predicate"
        "401":
          $ref: "#/components/responses/ApiKeyRequired"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
      description: Operator-minted `ik_...` API key.
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
      description: Page size. Values above 100 are clamped to 100.
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Row offset. Negative or invalid values become 0.
...
  responses:
...
    RateLimited:
      description: Request rate limit exceeded.
      headers:
        retry-after:
          schema:
            type: string
          description: Seconds until the bucket resets.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            rateLimited:
              value:
                error: rate_limited
    InternalError:
      description: Unhandled server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            internal:
              value:
                error: internal_error
  schemas:
...
    KgSchemaMetadata:
      type: object
      required: [schema, generatedAt, tables]
      properties:
        schema:
          type: string
          const: kg
        generatedAt:
          type: string
          format: date-time
        tables:
          type: array
          items:
            $ref: "#/components/schemas/SchemaTable"
...